Add OAuth2 login with success handler, refactor dependency injection, and update assets
- Introduced `OAuth2LoginSuccessHandler` to handle OAuth2 login flow and JWT token redirection.
- Updated `WebSecurityConfig` with OAuth2 login support and success handler.
- Added support for GitHub and Google OAuth2 providers.
- Refactored `UserService` and `UserDetailsServiceImpl` to use `@RequiredArgsConstructor` for cleaner dependency injection.
- Added Spring OAuth2 client dependency to `build.gradle.kts`.
- Updated favicon assets in the frontend.
