+++ razorlinks-web/index.html
@@ -2,7 +2,8 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/svg+xml" href="/images/img.svg" />
+      <link rel="icon" type="image/png" href="/images/img.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Razorlinks</title>
   </head>
