+++ razorlinks-web/src/components/LandingPage.jsx
@@ -72,7 +72,7 @@ const LandingPage = () => {
                         viewport={{ once: true }}
                         transition={{ duration: 0.8 }}
                         className="sm:w-[480px] w-[400px] object-cover rounded-md"
-                        src="/images/img2.png"
+                        src="/images/img.svg"
                         alt=""
                     />
                 </div>
